sparkle/requirements.txt
blendmodes
opencv-python
numpy
matplotlib
scikit-image
Pillow